HOTT-1776: Add support for separate error lib
